Update docker-compose.yml
Melisearch is a dependency for the frontend but isn't configured to be a healthy service before starting the web container.

Melisearch lacks a health check to ensure it operates as expected. Without a health check, it could run endlessly or restart without functioning properly.

Additionally, the web container lacks a defined timezone to output correct timestamps for debugging purposes.